refactor: move tarball extract to dedicated pkg
