+++ internal/tarball/suite_test.go
@@ -1,0 +1,20 @@
+package tarball_test
+
+import (
+	"testing"
+
+	. "gopkg.in/check.v1"
+
+	"github.com/canonical/chisel/internal/tarball"
+)
+
+func Test(t *testing.T) { TestingT(t) }
+
+type S struct{}
+
+var _ = Suite(&S{})
+
+func (s *S) SetUpTest(c *C) {
+	tarball.SetDebug(true)
+	tarball.SetLogger(c)
+}
